Refactor into single component per file.
