+++ airflow/ui/src/pages/Dashboard/HistoricalMetrics/MetricsBadge.tsx
@@ -16,19 +16,15 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { Box, Heading } from "@chakra-ui/react";
+import { Badge } from "@chakra-ui/react";
 
-import { Health } from "./Health";
-import { HistoricalMetrics } from "./HistoricalMetrics";
+type MetricBadgeProps = {
+  readonly color: string;
+  readonly runs: number;
+};
 
-export const Dashboard = () => (
-  <Box>
-    <Heading mb={4}>Welcome</Heading>
-    <Box>
-      <Health />
-    </Box>
-    <Box mt={5}>
-      <HistoricalMetrics />
-    </Box>
-  </Box>
+export const MetricsBadge = ({ color, runs }: MetricBadgeProps) => (
+  <Badge bg={color} borderRadius={15} minWidth={10} px={5} py={2} size="lg">
+    {runs}
+  </Badge>
 );
